+++ Assets/Scenes/SampleScene.unity
@@ -167,11 +167,16 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 388123cf76e9d784e8167644a85ad87d, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  weapon: 0
+  weapon: -1
   hp: 0
   totalSpeed: 0
   currentSpeed: 0
-  playerDeck: []
+  playerDeck:
+  - {fileID: 11400000, guid: 99b8ebdf2dba15440a3375d70f1039e9, type: 2}
+  - {fileID: 11400000, guid: 553adaf9f8bbb874c9bc3c6d0f01c9cf, type: 2}
+  - {fileID: 11400000, guid: 9b2525160dc7b524cbf56546f743960c, type: 2}
+  - {fileID: 11400000, guid: 7696040628e080b47b036bf524955b42, type: 2}
+  - {fileID: 11400000, guid: 9086ce301385e844fa55dbfcbf9dae1c, type: 2}
 --- !u!1 &378120923
 GameObject:
   m_ObjectHideFlags: 0
@@ -357,7 +362,7 @@ GameObject:
   - component: {fileID: 1117395797}
   - component: {fileID: 1117395798}
   m_Layer: 5
-  m_Name: Deck
+  m_Name: DefaultDeck
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -477,10 +482,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 80fd1647a6c5a0243921d584c5f98657, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  cards: []
-  cardsDatabase: []
+  playerDeck: []
   hand: {fileID: 1504785916}
   card: {fileID: 6344285022213485672, guid: b2f3be45166e6dd458ed65547b135076, type: 3}
+  player: {fileID: 327742870}
 --- !u!114 &1117395798
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -644,7 +649,7 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0.5, y: 0.5}
   m_AnchorMax: {x: 0.5, y: 0.5}
-  m_AnchoredPosition: {x: 0, y: -330}
+  m_AnchoredPosition: {x: 0, y: -430}
   m_SizeDelta: {x: 1920, y: 300}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1504785918
